Sonar tips

+++ src/creatures/players/player.cpp
@@ -2272,7 +2272,7 @@ void Player::removeExperience(uint64_t exp, bool sendText/* = false*/)
 		levelPercent = 0;
 	}
 	sendStats();
-	sendExperienceTracker(0, -exp);
+	sendExperienceTracker(0, -static_cast<int64_t>(exp));
 }
 
 double_t Player::getPercentLevel(uint64_t count, uint64_t nextLevelCount)
